Make setup script more robust and idempotent (#204)
* Make setup more robust and idempotent

While doing some development work for a different feature, I kept
running into issues with `make setup`. I'd run it after rebooting my
laptop and it would fail, or I'd run it and receive a weird error that I
wouldn't be able to reproduce. To address this, I've added more robust
error handling and have taken the opportunity to make the script
idempotent. You should be able to run `make setup` on a fully running
system and have nothing happen except for the API server and client
containers roll.

* Use editorconfig instead of vim comment

---------

Co-authored-by: Evan Foster <efoster@adobe.com>